Fix: Fix media paths not being replaced with relative paths
This bug affects all prior versions. Because media paths were not normalized to only use front slashes, and attachment (a type of media) used a capitalized drive letter, the subsequent replacement of absolute media paths with relative media paths fails. The user sees media with absolute paths in the final markdown, and the media is not rendered by markdown previewers because the `file://` uri scheme is not present in the URI.

The key is to normalize all media paths in the final markdown to use front slashes and with a lowercase drive letter, so that the search and replacement of absolute paths with relative paths in media paths works. Now the final markdown correctly contains relative media paths, and media is recognized in markdown previewers.