Fix: Add normalization of the notes destination path to have no trailing slashes or backslashes
When a user enters one or more trailing front slashes in the notes destination path, the script might error out on some Powershell versions due to invalid paths.

This fix adds normalization of the notes destination path to have no trailing slashes or backslashes. It does not change any functionality, but ensures that the script does not error out due to invalid paths.