Revert (configuration): Remove option to choose whether to keep spaces in file and folder names
Previously in #36, a configuration option `$keepPathSpaces` was added to control whether to keep spaces in section group, section, and page names in the generated files and folders.

It was identified in #53, that this configuration option will result in an error in pandoc, and has not been working since `v2.10.0`.

Pandoc throws errors when a path with spaces is given  in `-i` or `-o`. A possible workaround would be to normalize those arguments to use dashes instead of spaces, and then rename the final `-o` `.md` file to the original name with spaces.

Although it appears trivial to use the above workaround, it does mean that there would have to be two sets of output directories, one for the `.md` file output by `-o`, and the other for the final renamed file. This introduces complexity in the Page Conversion Configuration object. This is made more complex when the `$medialocation = 2` (where media is created in the same directory as the output `.md` file), such that the media must also be moved to the new location. Because media paths are not part of the Conversion Configuration, because media extraction is done by Pandoc, to move media  to their original location and to replace occurrences of any media path references in the `.md` file content must be done at conversion runtime. This has downsides:
- pollution of the conversion function
- difficult to test

In general, as discussed in https://github.com/theohbrothers/ConvertOneNote2MarkDown/issues/20#issuecomment-888956344, usage of space should be avoided, for the sake of 1) spaces in URIs are actually invalid 2) markdown compatibility issues with spaces in paths 3) A page named '  foo  bar  baz  ' should be expected to retain all spaces, but the expected filename '  foo bar baz  ' is actually invalid, so it would have to be trimmed to be 'foo bar baz', losing essential information in its name. Whereas if we did normalize that page name to use dashes, it would be '--foo-bar-baz--', retaining the whitespaces in the form of '-' and information in its name.

Hence, the default behavior would best be to normalize all spaces in Section Group, Section, and Page names to dashes.