Feature (configuration): Add option to choose whether to do a dry run
A new configuration option `dryRun` now allows to choose whether to do a conversion dry run. Use this in conjuction with `-Verbose` to get a very detailed overall sense of how the converted notes tree structure. This is useful for trying out between config options.

Once happy with configuration, ensure to turn `dryRun` off to run the actual conversion.