Enhancement: Use hierarchy-qualified static unique name for media
In all previous versions, media was named in format `<page-name><optional-postfix>-image<image-index>-<unique-timestamp><extension>`, e.g. `hello-world-image1-11223344.png`, and `hello-world-N-image1-11223344.png` for identically-named pages where `N` is incremented. This resulted in unique names for images, regardless of the setting of the `medialocation` configuration option.

However, if a user uses `$medialocation = 1` (i.e. a central `/media` folder), and converts a notebook with many pages with many images, if the notebook happens to contain many similarly-named or identically-named pages, the central `/media` folder can contain a lot of similarly named image names, e.g. `hello-world-image1-11223344.png`, `hello-world-1-1-image1-11223344.png`, where it becomes very difficult to make meaning of or trace a image to the page it belongs.

Now, the image names are named in format: `<hierarchy>-<page-name>-<optional-postfix>-image<image-index>.<extension`. The difference is that instead of using a timestamp suffix, a hierarchy prefix is added. This gives meaning to the image name, while preserving the uniqueness of all image names within the scope of a notebook.

Another benefit of this is that repeated conversion of a notebook can be done against an existing output folder, without creating duplicate images (which was the previous behavior when using a timestamp suffix).

Note that because media is generated at conversion runtime, this feature cannot really be unit tested. The code is straightforward since the naming of the image makes use of a conversion configuration object.