Feature (configuration): Add configuration option `newlineCharacter` to choose whether to use LF or CRLF for newlines
This feature enables a user to choose whether to use Line Feed (`LF`) or Carriage Return + Line Feed (`CRLF`) for newlines in `.md` files. This is especially useful when performing conversion against a notes destination folder which is `git` controlled, such that the generated `.md` do not appear in `git diff` as a result of newline diffs.

The default behavior is to use `LF`.