Fix: Fix inserted attachment names not being markdown-encoded in final markdown
This bug affects all previous versions.

Previously, if an attachment name contains any of the following characters:

```
*_{}[]()#+-.!`
```

then the reference of the attachment in markdown becomes unrecognized by markdown previewers because those characters are not markdown-escaped (i.e. interpreted as markdown) when the name of the reference is injected into the markdown during the conversion.

Now, these characters are markdown-encoded, and references of attachments are now valid and interpretable by markdown previewers.