Enhancement: Improve conversion speed for notebooks containing many media files
Previously, when renaming files to unique names, conversion function looked in the `/media` directory for images (`Get-ChildItem`) with basename starting with `image`. However, as the number of files in the `/media` grows, conversion slows.

Now, pandoc converts is made to extract media into a unique temp media folder, and the conversion function looks in that folder for images with basename starting with `image`. This temp media folder will only ever contain the media specific to the converted page, and hence `Get-ChildItem` stays performant.

This temp folder will be empty at each conversion, and will not be deleted, for safety reasons (`Remove-Item -Recurse` is a dangerous command).

In addition, `-Verbose` now shows markdown mutation's regexes and replacements.