Enhancement: Remove isolated occurrences of '>' in bullet lists
It is very common to have an "empty bullet point" in a onenote bullet list. This results in the conversion leaving isolated `>` in markdown.

E.g.

```
My list:
- foo

- bar

>

hello world
```

`>` in markdown represents a quote, and so this element should be removed.

Now, this mutation is part of the configuration option `$keepspaces = 1` (default behavior)