Fix: Fix broken paths for second level page preceded by a third level page preceded by a first level page
Fixes #96

The scenario of 'second level page preceded by a third level page preceded by a first level page' can cause a double `Split-Path -Parent` which should only involve a single `Split-Path -Parent`. Because `['pagePrefix']` ended up with a value of `\`, when `[io.path]::combine` involves any path that begins with `\`, the path immediately begins with `\`, which is an invalid path in Windows.

The determination of the `['pagePrefix']` is now fixed for this particular scenario of 'second level page preceded by a third level page preceded by a first level page'. Unit tests have been added so this fix should work reliably.