Enhancement (configuration): Add option `mdFileNameAndFolderNameMaxLength` to limit markdown file name and parent folder names to avoid hitting filesystem limits of 255 bytes
Previously, the default behavior was to normalize all markdown file and parent folders to a maximum of 255 characters. Based on https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits, the file and folder name limit is 255 bytes on most filesystems. However, in certain cases, when special multibyte characters are present in the file or folder names, even normalizing file and folder names to 255 characters might exceed 255 bytes. This results in failed conversions.

Now, a new configuration option `mdFileNameAndFolderNameMaxLength` is added to allow the setting of a limit on the file and folder names, to avoid hitting the filesystem limits of 255 bytes for file and folder names. The default value is 32 characters.

A lower value can help avoid hitting [file and folder name limits of `255` bytes on file systems](https://en.wikipedia.org/wiki/Comparison_of_file_systems#Limits). A higher value preserves a longer title. If using page prefixes (e.g. `Page_Subpage.md`), it is recommended to use a value of `100` or greater.