Feature (configuration): Add `exportPdf` option to include a `.pdf` export alongside `.md` file
Currently, `InkDrawing` (i.e. highlights, pen drawings) absolute positioning of a page is not preserved in the intermediate `.docx`, and hence also not preserved in the final `.md`. One easy solution to preserve the entire state of the page is to use a document that supports absolute positioning - e.g. HTML, PDF. However, HTML export is not an singular HTML file, since images are exported as separate files from the `.htm`.

Hence, a new configuration option `exportPdf` is added, which allows the user to specify whether to include a `.pdf` alongside each `.md` file. `.md` does not preserve `InkDrawing` absolute positions, but a `.pdf` export is a complete page snapshot that also preserves `InkDrawing` positions.