Enhancement (config): Change `dryRun` config option to a boolean integer
Previously, `dryRun` config option used integer values `1` and `2`, in order to mirror other non-boolean config options. While it may not be a problem for non-technical users, technical users may find that it makes no sense to have `$dryRun = 1` do an actual conversion, and `$dryRun = 2` do a dry run. It has become apparent that `dryRun` is not more intuitive being an integer.

Now, `dryRun` is changed to a boolean (`0` or `1`). Retaining its integer type keeps it easy to use, rather than using actual boolean values `$true` and `$false` which are difficult to type and error prone. This keeps the `dryRun` config options easy to use for both non-technical users and technical users by not violating the principle of least astonishment. Simply set `$dryRun = 0` to disable dry run, and `$dryRun = 1` to enable dry run.