Docs (readme): Add notes about Windows Powershell being included Windows 10 and 11, and not to use Windows Powershell ISE
In #125, it was thought that Windows Powershell 5.x on Windows 11 did not support long path using the `\\?\` prefix, e.g. `\\?\C:\path\to`. However, further investigation identified that Windows Powershell ISE did not support the `\\?\` prefix, but Windows Powershell did.

Now, Windows Powershell 5.x is known to be fully compatible with this script. There is not need to install Powershell Core, since Windows 10 and 11 already include Windows Powershell. A note is added in requirements section.

Closes #125